removed email verification for demo

+++ Backend/src/controllers/auth.controller.js
@@ -23,6 +23,10 @@ export const signup = async (req, res) => {
     if (!user) {
       user = await User.create({ name, email, password });
     }
+    if (process.env.ENABLE_EMAIL_VERIFICATION !== "true") {
+      user.isVerified = true;
+      await user.save();
+    }
 
     await Otp.deleteMany({ email, purpose: "SIGNUP" });
 
@@ -44,6 +48,7 @@ export const signup = async (req, res) => {
     res.status(201).json({
       success: true,
       message: "Signup successful. OTP sent to your email.",
+      isVerified: user.isVerified,
     });
   } catch (error) {
     console.error(error);
@@ -61,7 +66,7 @@ export const verifyOtp = async (req, res) => {
       return res.status(400).json({ message: "Invalid or expired OTP" });
     }
 
-    // Compare Date objects 
+    // Compare Date objects
     if (new Date(otpRecord.expiresAt) < new Date()) {
       return res.status(400).json({ message: "Invalid or expired OTP" });
     }
@@ -111,15 +116,15 @@ export const login = async (req, res) => {
       return res.status(401).json({ message: "Invalid email or password" });
     }
 
-    // Check verification AFTER password match
-    if (!user.isVerified) {
-      // Remove old signup OTPs
+    // Disabled resend email for sometime
+    const emailVerificationEnabled =
+      process.env.ENABLE_EMAIL_VERIFICATION === "true";
+
+    if (!user.isVerified && emailVerificationEnabled) {
       await Otp.deleteMany({ email, purpose: "SIGNUP" });
 
-      // Generate OTP
       const otp = generateOtp();
 
-      // Store OTP in Otp collection
       await Otp.create({
         email,
         otp,
@@ -127,7 +132,6 @@ export const login = async (req, res) => {
         expiresAt: new Date(Date.now() + 5 * 60 * 1000),
       });
 
-      // Send email
       await sendEmail({
         to: email,
         subject: "Verify your Taskora account",
@@ -149,7 +153,7 @@ export const login = async (req, res) => {
         isRunner: user.isRunner,
       },
       process.env.JWT_SECRET,
-      { expiresIn: "7d" }
+      { expiresIn: "7d" },
     );
 
     // Return user data (without password)
@@ -167,7 +171,6 @@ export const login = async (req, res) => {
       message: "Login successful",
       user: userData,
       token,
-      
     });
   } catch (error) {
     console.error(error);
